refactor(rust): assemble the tonic channel in armonik, not in the transport
`armonik-transport` hands back a connector; wrapping it in an HTTP/2 engine is the
consumer's business. `connect` moves to `armonik/src/client/channel.rs`, a module of
its own next to the client: it is the only thing in this workspace that turns a
connector into a channel, and naming it after what it produces keeps `client/mod.rs`
about the services.

Interpreting `Endpoint` stays here, as `impl TryFrom<&HttpConfig> for http::Uri`: the
URI a request is addressed to, which is the endpoint unless `OverrideTargetName` moves
it, and which refuses an unset endpoint by the option's name. Both are this crate's
vocabulary, and one conversion is the whole of it - there is no second public name for
resolving a configuration.

That URI is resolved once per connection. `https_connector` takes it rather than
deriving it again: `channel.rs` needs it for the channel's origin and the connector
needs it for the name the certificate is verified against, and having the connector
re-derive it made the caller resolve twice. Taking it, rather than returning it
alongside the connector, keeps `https_connector`'s return type a connector and puts the
resolution where the reader sees it, at the top of the caller. `TlsConfig::override_target`
stays `pub(crate)`: the conversion is its only caller.

Whether to pin the name the certificate is verified against is decided on
`OverrideTargetName` being set, rather than on an `Option` the resolution handed over;
the two are the same condition. The `ServerName::try_from(...).expect(...)` under it
still panics on an override written as a bracketed IPv6 literal such as `[::1]`,
unchanged from before this move: the guard narrows the window to the no-override case
rather than closing it. Closing it is a fix of its own.

The error follows the code. Transport's `ConnectionError` sheds its `Config` variant,
which only the resolution built, and reports what building a connector can fail at:
TLS and the system cert store. `armonik::client::ChannelError` carries the
configuration refusal, the connector's own failure and the tonic transport failure;
`Client::with_config` and `NewClientError::Connect` name it.

The `Timeout`, `RateLimit`, `Http2*` and `UserAgent` docs move to the model the `retry`
field already uses: they name who applies the option rather than claim a function that
no longer lives here reads it.

The integration tests that drive a real channel (`timeout`, `proxy`, and the slow gRPC
service they share) move with `connect`, since `armonik-transport` can no longer build
a channel to call through. `armonik-transport` keeps no `tonic` dev-dependency.

Verified from packages/rust:

  cargo build --workspace --all-features            ok
  cargo clippy --workspace --all-targets --all-features -- -D warnings   clean
  cargo clippy --workspace --all-features --no-deps -- -Dwarnings -Dunused-crate-dependencies
    clean, which is what CI runs
  cargo fmt --check                                 clean
  RUSTDOCFLAGS=-Dwarnings cargo doc --workspace --no-deps --all-features  clean
  cargo test --workspace --all-features --no-fail-fast

    Every target passes but one, and that one is the known exception: 110 of
    `armonik`'s lib tests reach the ArmoniK.Api.Mock server through `get_nb_request`,
    which is not running here. The same 110 fail on this commit's base, so the count is
    unmoved.

    What this commit is answerable for: `armonik-transport`'s lib tests lose the
    connector's empty-endpoint test and gain three for the `Uri` conversion, a delta of
    +2 with nothing dropped; `armonik` gains the `proxy` and `timeout` targets, 15 and 6
    tests moved here whole, and one lib test, `channel`.

    Suite totals are not quoted as properties of a crate: they count every commit
    underneath this one and move whenever the stack is reordered. As one measurement of
    this commit where it currently sits, `armonik-transport`'s lib runs 106, all green.
