+++ packages/rust/armonik-transport/Cargo.toml
@@ -28,9 +28,9 @@ env = ["serde"]
 schema = ["serde", "dep:schemars", "dep:serde_json"]
 
 [dependencies]
-# `channel` for `tonic::transport::Endpoint`, which is what `connect` builds; `codegen` for the
-# `http`/`tokio_stream` re-exports a dependent reaches through this crate.
-tonic = { workspace = true, features = ["channel", "codegen"] }
+# Only re-exported, so a dependent assembling a channel reaches the version the workspace pins.
+# Nothing here builds against it.
+tonic.workspace = true
 snafu.workspace = true
 tracing.workspace = true
 hyper = { workspace = true, features = ["client", "http1", "http2"] }
@@ -86,9 +86,6 @@ tokio = { workspace = true, features = ["net", "time"] }
 tower-service.workspace = true
 
 [dev-dependencies]
-# The proxy tests that follow the environment set process-global variables, hence one test at a
-# time.
-serial_test.workspace = true
 # The unit tests build configurations the way a caller does under the `serde` feature: from a JSON
 # document naming the flat options.
 serde_json.workspace = true
@@ -97,12 +94,6 @@ serde_json.workspace = true
 # really signed the leaf, which a committed fixture would also have to keep unexpired.
 rcgen.workspace = true
 tokio = { workspace = true, features = ["rt-multi-thread", "macros", "time"] }
-# The integration tests serve a gRPC service to call, which the regular build never does; these features
-# union with the `channel`/`codegen` above for test builds only.
-tonic = { workspace = true, features = ["server", "router"] }
-# The test service's codec moves raw `Bytes`; production code here hands whole channels to its caller
-# rather than framing messages itself.
-bytes.workspace = true
 
 [[example]]
 name = "generate_schema"
